streaming/dashboard/requirements.txt
fastapi
uvicorn[standard]
ssi-fc-data  # nếu bạn dùng thư viện này
kafka-python
yfinance 
vnstock
confluent-kafka
orjson